requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0

django-oauth-toolkit==1.0.0
oauthlib==2.0.2

# Bleeding edge Django
django==1.10.8 # pyup: >=1.10,<1.11

# Configuration
django-environ==0.4.4
whitenoise==3.3.1


# Forms
django-crispy-forms==1.7.0

# Models
django-model-utils==3.0.0

# Images
Pillow==4.3.0

# Password storage
argon2-cffi==16.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.34.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.3.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.3

# Redis support
django-redis==4.8.0
redis>=2.10.5


# Rest Framework / API / Sockets
djangorestframework==3.7.3
djangorestframework_jwt==1.11.0
djangorestframework-recursive==0.1.2

markdown==2.6.9
django_filter==1.1.0
django_extensions==1.9.9


rcssmin==1.0.6
django-compressor==2.2
django-import-export==0.6.0

# Your custom requirements go here
git+https://github.com/jet-community/django-jet.git@master
git+https://github.com/millerhooks/django-jsoneditor.git
django_guardian==1.4.9
